requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
future==1.0.0
idna==3.6
iso8601==2.1.0
MouseInfo==0.1.3
pillow==10.2.0
PyAudio==0.2.14
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pytweening==1.2.0
pywin32==306
PyYAML==6.0.1
requests==2.31.0
SpeechRecognition==3.10.1
typing_extensions==4.10.0
urllib3==2.2.1
